mcp-servers/openai-judge/pyproject.toml
[project]
name = "openai-judge-mcp"
version = "0.1.0"
description = "MCP server for OpenAI-based code review as LLM Judge"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.1.2",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
